+++ TabletFriend/TabletFriend/LayoutChangeMethod.cs
@@ -1,0 +1,8 @@
+﻿namespace TabletFriend
+{
+	public enum LayoutChangeMethod
+	{
+		Manual,
+		Automatic
+	}
+}